Move WorkspacePanel to IDE
