+++ src/ui/ide/WorkspacePanel/index.css
