+++ src/ui/ide/WorkspacePanel/useBlocklyWorkspace.ts
@@ -1,6 +1,6 @@
 import {useEffect, useRef} from "react";
 import * as Blockly from "blockly";
-import {injectWorkspace, setupWorkspace} from "../../core/blockly";
+import {injectWorkspace, setupWorkspace} from "../../../core/blockly";
 
 export default function useBlocklyWorkspace() {
   const divRef = useRef<HTMLDivElement>(null)
