--- src/ui/workspace/index.ts
@@ -1,1 +1,0 @@
-export {default as WorkspacePanel} from './WorkspacePanel.tsx'
